pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Safiullin, Ruslan <rus.safiullin.01@mail.ru>"]
readme = "README.md"
packages = [{ include = "page_analyzer" }]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.0"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.0"
beautifulsoup4 = "^4.12.2"
requests = "^2.31.0"
validators = "^0.22.0"
gunicorn = "^21.2.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


# psycopg2-binary python-dotenv beautifulsoup4 requests validators gunicorn
